requirements.txt
transformers
torch
fastapi
uvicorn
pillow
python-multipart
